[ISSUE-2] Support CocoaPods Analyzer
