+++ lib/fastlane/plugin/dependency_check_ios_analyzer/version.rb
@@ -1,5 +1,5 @@
 module Fastlane
   module DependencyCheckIosAnalyzer
-    VERSION = '0.3.0'
+    VERSION = '1.0.0'
   end
 end
