+++ sample-project/Podfile
@@ -1,0 +1,6 @@
+platform :ios, '11.0'
+
+target 'sample-project' do
+    pod 'AFNetworking', '= 2.5.2'
+    pod 'SwiftyJSON', '~> 4.0'
+end
